requirements.txt
Django==5.1.1
django_phonenumber_field==8.1.0
djangorestframework_simplejwt==5.5.0
python-decouple==3.8
